Tidy up packages for the Scala challenges in the Americium entry.
